requirements.txt
pip~=24.0
setuptools~=69.2.0
wheel~=0.43.0
virtualenv

omegaconf~=2.3.0
cachetools~=5.3.3
loguru~=0.7.2
watchdog~=4.0.0
humanfriendly

numpy
pandas
matplotlib

aiohttp

alpaca-py
backtrader


pyfolio